Fix help_text logic and allow it to contain links
there was a lot of special logic to allow some implementations to include links in the help_text. However, this had a lot of code duplicate and barely worked. For example, it only worked for multi_help.
Now a macro exists that can convert a regular string, containing <a></a> tags, to html with these tags rendered as links.

`multi_help` variable is also removed: the template can just render the help_text inside `choices` if it exists. No need to have an extra variable for it.