align look of multiple_checkbox_compliance regardles of help text location
For some reason this file had 2 looks originally. Looking at it closer, this seems to be a bug: the look without the styling doesn't work correctly. It was probably fixed for one instance (when using multiple help texts), but not for the default case.

This aligns those cases to just always show the correct, working label and checkbox input.